Refactor marketplace callbacks into shared ACL
