+++ shared-lib/shared-common/src/main/java/com/ejada/common/marketplace/dto/EnvironmentIdentifierDto.java
@@ -1,0 +1,9 @@
+package com.ejada.common.marketplace.dto;
+
+/**
+ * Identifier values we may return to the marketplace after provisioning completes.
+ */
+public record EnvironmentIdentifierDto(
+        String identifierCd,
+        String identifierValue) {
+}
