+++ shared-lib/shared-common/src/main/java/com/ejada/common/marketplace/dto/SubscriptionFeatureDto.java
@@ -1,0 +1,11 @@
+package com.ejada.common.marketplace.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+/**
+ * Feature entitlements provisioned with the subscription.
+ */
+public record SubscriptionFeatureDto(
+        @NotBlank String featureCd,
+        Integer featureCount) {
+}
