+++ shared-lib/shared-common/src/main/java/com/ejada/common/marketplace/dto/SubscriptionUpdateType.java
@@ -1,0 +1,11 @@
+package com.ejada.common.marketplace.dto;
+
+/**
+ * Update events emitted by the marketplace to drive subscription lifecycle transitions.
+ */
+public enum SubscriptionUpdateType {
+    SUSPENDED,
+    RESUMED,
+    TERMINATED,
+    EXPIRED
+}
