+++ tenant-platform/billing-service/src/main/java/com/ejada/billing/service/impl/ConsumptionServiceImpl.java
@@ -11,6 +11,8 @@ import com.ejada.billing.model.UsageCounter;
 import com.ejada.billing.repository.UsageCounterRepository;
 import com.ejada.billing.repository.UsageEventRepository;
 import com.ejada.billing.service.ConsumptionService;
+import com.ejada.common.dto.ServiceResult;
+import com.ejada.common.security.TokenHashUtils;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.dao.DataIntegrityViolationException;
@@ -21,12 +23,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.transaction.interceptor.TransactionAspectSupport;
 import org.springframework.transaction.support.TransactionTemplate;
 
-import com.ejada.common.dto.ServiceResult;
-
 import java.math.BigDecimal;
-import java.nio.charset.StandardCharsets;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
@@ -133,7 +130,7 @@ public class ConsumptionServiceImpl implements ConsumptionService {
                                      final TrackProductConsumptionRq rq) {
         eventRepo.save(eventMapper.build(
                 rqUid,
-                sha256(token),
+                TokenHashUtils.sha256(token),
                 toJson(rq),
                 rq != null ? rq.productId() : null,
                 "I000000",
@@ -162,7 +159,7 @@ public class ConsumptionServiceImpl implements ConsumptionService {
             requiresNewTx.executeWithoutResult(status ->
                     eventRepo.save(eventMapper.build(
                             rqUid,
-                            sha256(token),
+                            TokenHashUtils.sha256(token),
                             safeJson(rq),
                             rq != null ? rq.productId() : null,
                             "EINT000",
@@ -190,20 +187,4 @@ public class ConsumptionServiceImpl implements ConsumptionService {
         }
     }
 
-    private String sha256(final String s) {
-        if (s == null) {
-            return null;
-        }
-        try {
-            MessageDigest md = MessageDigest.getInstance("SHA-256");
-            byte[] b = md.digest(s.getBytes(StandardCharsets.UTF_8));
-            StringBuilder sb = new StringBuilder(b.length * 2);
-            for (byte x : b) {
-                sb.append(String.format("%02x", x));
-            }
-            return sb.toString();
-        } catch (NoSuchAlgorithmException e) {
-            return null;
-        }
-    }
 }
