--- tenant-platform/subscription-service/src/main/java/com/ejada/subscription/dto/EnvironmentIdentifierDto.java
@@ -1,8 +1,0 @@
-package com.ejada.subscription.dto;
-
-
-/** Identifier values we may return to marketplace after provisioning. */
-public record EnvironmentIdentifierDto(
-     String identifierCd,     // e.g., DB_ID
-     String identifierValue   // e.g., "10.20.0.1"
-) { }
