--- tenant-platform/subscription-service/src/main/java/com/ejada/subscription/dto/ProductPropertyDto.java
@@ -1,8 +1,0 @@
-package com.ejada.subscription.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record ProductPropertyDto(
-    @NotBlank String propertyCd,
-    @NotBlank String propertyValue
-) { }
