--- tenant-platform/subscription-service/src/main/java/com/ejada/subscription/dto/SubscriptionFeatureDto.java
@@ -1,8 +1,0 @@
-package com.ejada.subscription.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record SubscriptionFeatureDto(
-    @NotBlank String featureCd,
-    Integer featureCount
-) { }
