+++ tenant-platform/subscription-service/src/main/java/com/ejada/subscription/service/SubscriptionInboundService.java
@@ -1,9 +1,9 @@
 package com.ejada.subscription.service;
 
-import com.ejada.subscription.dto.ReceiveSubscriptionNotificationRq;
-import com.ejada.subscription.dto.ReceiveSubscriptionNotificationRs;
-import com.ejada.subscription.dto.ReceiveSubscriptionUpdateRq;
 import com.ejada.common.dto.ServiceResult;
+import com.ejada.common.marketplace.dto.ReceiveSubscriptionNotificationRq;
+import com.ejada.common.marketplace.dto.ReceiveSubscriptionNotificationRs;
+import com.ejada.common.marketplace.dto.ReceiveSubscriptionUpdateRq;
 
 import java.util.UUID;
